+++ src/official-widgets/icon-triggered-grid/icon-triggered-grid.spec.tsx
@@ -22,7 +22,7 @@ describe('icon-triggered-grid', () => {
     en: {
       widgetTitle: 'Icon Triggered Grid 814',
       errorDescription: 'Houston, we\'ve had a problem!',
-      triggerCTA: 'Find Similar 123',
+      triggerCTA: 'CTA',
     },
   };
   const mockVisearchClient: ViSearchClient = {
@@ -135,7 +135,8 @@ describe('icon-triggered-grid', () => {
         expect(pid).toBe('pid-not-found');
         expect(params).toEqual({
           return_product_info: true,
-          show_best_product_images: true,
+          limit: 20,
+          show_best_product_images: false,
           sort_by: '',
           facets: [
             'price',
@@ -186,7 +187,8 @@ describe('icon-triggered-grid', () => {
         expect(pid).toBe('pid-not-found');
         expect(params).toEqual({
           return_product_info: true,
-          show_best_product_images: true,
+          limit: 20,
+          show_best_product_images: false,
           sort_by: '',
           facets: [
             'price',
@@ -227,7 +229,8 @@ describe('icon-triggered-grid', () => {
         expect(pid).toBe('pid-found');
         expect(params).toEqual({
           return_product_info: true,
-          show_best_product_images: true,
+          limit: 20,
+          show_best_product_images: false,
           sort_by: '',
           facets: [
             'price',