+++ src/official-widgets/more-like-this/more-like-this.spec.tsx
@@ -79,7 +79,8 @@ describe('more-like-this', () => {
         expect(pid).toBe('pid-not-found');
         expect(params).toEqual({
           return_product_info: true,
-          show_best_product_images: true,
+          limit: 20,
+          show_best_product_images: false,
           sort_by: '',
           facets: [
             'price',
@@ -111,7 +112,8 @@ describe('more-like-this', () => {
         expect(pid).toBe('pid-found');
         expect(params).toEqual({
           return_product_info: true,
-          show_best_product_images: true,
+          limit: 20,
+          show_best_product_images: false,
           sort_by: '',
           facets: [
             'price',