+++ src/official-widgets/similar-search/dev-configs.ts
@@ -10,10 +10,6 @@ export const devConfigs: RecursivePartial<WidgetConfig> = {
     placementId: '',
     endpoint: '',
   },
-  searchSettings: {
-    facets_limit: 10,
-    limit: 20,
-  },
   displaySettings: {
     cssSelector: '.similar-search-widget',
     productDetails: {},